--- Project Bloodfield/Assets/Art/Alpine/weapons/IceShard.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0c25069368875b64688491f51cfaf6bd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
