--- Project Bloodfield/Assets/Art/Alpine/weapons/IceShard/IceShardTRavel.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0dcc08c25cf9acb41ae375dfdeacc503
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
