+++ Project Bloodfield/Assets/Art/Alpine/weapons/Player-Spells.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: abaf56ad88ee3a84b9289fb974479dd5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
