+++ Project Bloodfield/Assets/Art/Alpine/weapons/Weapon-Models.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06aea0df919d72644aa32fa2d6b5b980
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
