+++ Project Bloodfield/Assets/Prefabs/IceShard/DamageText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af0e01ccbd2ceec408aa13d749b88f30
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
