+++ Project Bloodfield/Assets/Prefabs/IceShard/Ice-shard-bullet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41b74e1f9cc0a5c43b5eb23c4ff27225
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
