+++ Project Bloodfield/Assets/Prefabs/IceShard/IceShard.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bce045651e5b48441babf253610f81e0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
