+++ Project Bloodfield/Assets/Prefabs/IceShard/IceShard.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 442ba143a99f4ad43a64588623ceacb8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
