+++ Project Bloodfield/Assets/Prefabs/IceShard/IceShardExplosion.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee12825b0a8a61e41bfe6a4735965595
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
