+++ Project Bloodfield/Assets/Prefabs/IceShard/IceShardTraveling.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65f60a4cb4a1f6948959dac903cfbb7f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
