+++ Project Bloodfield/Assets/Prefabs/IceShard/Player_Shooting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c743baf534633fe458eb897603d0a78c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
