--- Project Bloodfield/Assets/Scripts/Player-Spells.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: abaf56ad88ee3a84b9289fb974479dd5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
