--- Project Bloodfield/Assets/Scripts/Player-Spells/Ice-Shard.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7887be48507b8344d922500bc2ace9df
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
